+++ src/cli.ts
@@ -187,7 +187,7 @@ type ValidateOptions = {
    * Whether to validate that each release section has one or more changelog
    * entries (true) or not (false).
    */
-  noEmptyReleases: boolean;
+  disallowEmptyReleases: boolean;
 };
 
 /**
@@ -205,7 +205,7 @@ type ValidateOptions = {
  * @param options.ensureValidPrLinksPresent - Whether to validate that each
  * changelog entry has one or more links to associated pull requests within the
  * repository (true) or not (false).
- * @param options.noEmptyReleases - Whether to validate that each release
+ * @param options.disallowEmptyReleases - Whether to validate that each release
  * section has one or more changelog entries (true) or not (false).
  */
 async function validate({
@@ -218,7 +218,7 @@ async function validate({
   formatter,
   packageRename,
   ensureValidPrLinksPresent,
-  noEmptyReleases,
+  disallowEmptyReleases,
 }: ValidateOptions) {
   const changelogContent = await readChangelog(changelogPath);
 
@@ -232,7 +232,7 @@ async function validate({
       formatter,
       packageRename,
       ensureValidPrLinksPresent,
-      noEmptyReleases,
+      disallowEmptyReleases,
     });
     return undefined;
   } catch (error) {
@@ -412,8 +412,8 @@ async function main() {
               'Verify that each changelog entry has one or more links to associated pull requests within the repository',
             type: 'boolean',
           })
-          .option('noEmptyReleases', {
-            default: false,
+          .option('disallowEmptyReleases', {
+            default: true,
             description:
               'Verify that each release section has one or more changelog entries',
             type: 'boolean',
@@ -442,7 +442,7 @@ async function main() {
     tagPrefixBeforePackageRename,
     autoCategorize,
     prLinks,
-    noEmptyReleases,
+    disallowEmptyReleases,
     useChangelogEntry,
     useShortPrLink,
     requirePrNumbers,
@@ -597,7 +597,7 @@ async function main() {
       formatter,
       packageRename,
       ensureValidPrLinksPresent: prLinks,
-      noEmptyReleases,
+      disallowEmptyReleases: Boolean(disallowEmptyReleases),
     });
   } else if (command === 'init') {
     await init({