+++ src/validate-changelog.test.ts
@@ -515,7 +515,7 @@ describe('validateChangelog', () => {
       ).resolves.not.toThrow();
     });
 
-    it('should not throw if the changelog has an empty release', async () => {
+    it('should not throw if the changelog has an empty release and disallowEmptyReleases is false', async () => {
       const changelogWithEmptyRelease = changelogWithReleases.replace(
         '## [1.0.0] - 2020-01-01\n### Changed\n- Something else\n',
         '## [1.0.0] - 2020-01-01\n',
@@ -527,11 +527,12 @@ describe('validateChangelog', () => {
           repoUrl:
             'https://github.com/ExampleUsernameOrOrganization/ExampleRepository',
           isReleaseCandidate: false,
+          disallowEmptyReleases: false,
         }),
       ).resolves.not.toThrow();
     });
 
-    it('should throw if the changelog has an empty release and noEmptyReleases is true', async () => {
+    it('should throw if the changelog has an empty release and disallowEmptyReleases is true', async () => {
       const changelogWithEmptyRelease = changelogWithReleases.replace(
         '## [1.0.0] - 2020-01-01\n### Changed\n- Something else\n',
         '## [1.0.0] - 2020-01-01\n',
@@ -543,7 +544,23 @@ describe('validateChangelog', () => {
           repoUrl:
             'https://github.com/ExampleUsernameOrOrganization/ExampleRepository',
           isReleaseCandidate: false,
-          noEmptyReleases: true,
+          disallowEmptyReleases: true,
+        }),
+      ).rejects.toThrow("Release has no changelog entries: '1.0.0'");
+    });
+
+    it('should throw by default if the changelog has an empty release', async () => {
+      const changelogWithEmptyRelease = changelogWithReleases.replace(
+        '## [1.0.0] - 2020-01-01\n### Changed\n- Something else\n',
+        '## [1.0.0] - 2020-01-01\n',
+      );
+      await expect(
+        validateChangelog({
+          changelogContent: changelogWithEmptyRelease,
+          currentVersion: '1.0.1',
+          repoUrl:
+            'https://github.com/ExampleUsernameOrOrganization/ExampleRepository',
+          isReleaseCandidate: false,
         }),
       ).rejects.toThrow("Release has no changelog entries: '1.0.0'");
     });
@@ -560,6 +577,7 @@ describe('validateChangelog', () => {
           repoUrl:
             'https://github.com/ExampleUsernameOrOrganization/ExampleRepository',
           isReleaseCandidate: false,
+          disallowEmptyReleases: false,
         }),
       ).rejects.toThrow('Changelog is not well-formatted');
     });
@@ -638,11 +656,12 @@ describe('validateChangelog', () => {
           repoUrl:
             'https://github.com/ExampleUsernameOrOrganization/ExampleRepository',
           isReleaseCandidate: true,
+          disallowEmptyReleases: false,
         }),
       ).resolves.not.toThrow();
     });
 
-    it('should throw if the changelog has an empty release and noEmptyReleases is true', async () => {
+    it('should throw if the changelog has an empty release and disallowEmptyReleases is true', async () => {
       const changelogWithEmptyRelease = changelogWithReleases.replace(
         '## [1.0.0] - 2020-01-01\n### Changed\n- Something else\n',
         '## [1.0.0] - 2020-01-01\n',
@@ -654,7 +673,23 @@ describe('validateChangelog', () => {
           repoUrl:
             'https://github.com/ExampleUsernameOrOrganization/ExampleRepository',
           isReleaseCandidate: true,
-          noEmptyReleases: true,
+          disallowEmptyReleases: true,
+        }),
+      ).rejects.toThrow("Release has no changelog entries: '1.0.0'");
+    });
+
+    it('should throw by default if the changelog has an empty release', async () => {
+      const changelogWithEmptyRelease = changelogWithReleases.replace(
+        '## [1.0.0] - 2020-01-01\n### Changed\n- Something else\n',
+        '## [1.0.0] - 2020-01-01\n',
+      );
+      await expect(
+        validateChangelog({
+          changelogContent: changelogWithEmptyRelease,
+          currentVersion: '1.0.0',
+          repoUrl:
+            'https://github.com/ExampleUsernameOrOrganization/ExampleRepository',
+          isReleaseCandidate: true,
         }),
       ).rejects.toThrow("Release has no changelog entries: '1.0.0'");
     });
@@ -671,6 +706,7 @@ describe('validateChangelog', () => {
           repoUrl:
             'https://github.com/ExampleUsernameOrOrganization/ExampleRepository',
           isReleaseCandidate: false,
+          disallowEmptyReleases: false,
         }),
       ).rejects.toThrow('Changelog is not well-formatted');
     });